requirements.txt
Flask
opencv-python
telethon
python-dotenv
